requirements.txt
# Local LLM dependencies
llama-cpp-python==0.2.79
huggingface-hub==0.23.0

# Web framework - FastAPI is modern, fast, and has great documentation
fastapi==0.109.0

# ASGI server - runs the FastAPI application
uvicorn==0.27.0

# Pydantic - data validation (used by FastAPI)
pydantic==2.5.3

# For async HTTP requests if needed
httpx==0.26.0
